pyproject.toml
[tool.poetry]
name = "overtrack-web-2"
version = "0.2.0"
description = ""
authors = ["Simon Pinfold <simon@uint8.me>"]

[tool.poetry.dependencies]
python = "^3.6"

overtrack-models = { git = "https://gitlab.com/OverTrack/overtrack-models.git", branch = "master" }

dataclasses = "*"

typedload = "*"
zappa = { git = "https://github.com/overtrack-gg/Zappa.git", branch = "master" }

flask = "^1"
bootstrap-flask = "*"
pyjwt = "*"
requests_oauthlib = "*"
shortuuid = "*"

pynamodb = "^4"

pytest = "^4.4"
datadog = "^0.29"
sentry_sdk = "^0.11"

stripe = "^2.35"

[tool.poetry.dev-dependencies]

#pytest = "*"
#pytest-ordering = "*"
#pytest-sugar = "*"
#pytest-cov = "*"
libsass = "^0.19"
boussole = "^1.5"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
